# src/main/resources/openapi/hict-api-v1.yaml
openapi: 3.0.3
info:
  title: HiCT JVM API
  version: v1
  description: |
    HTTP API for HiCT_JVM server.

    This specification is intended for WebUI integration, Python `hict`/`hict_jvm_api`
    clients and programmatic data-processing pipelines.
servers:
  - url: /
...
paths:
...
  /reverse_selection_range:
    post:
      tags: [Scaffolding]
      summary: Reverse orientation of selected assembly range
      operationId: reverseSelectionRange
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RangeRequest' }
      responses:
        '200':
          description: Updated assembly info with version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenericMap' }
  /move_selection_range:
    post:
      tags: [Scaffolding]
      summary: Move selected range to a new target position
      operationId: moveSelectionRange
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MoveSelectionRequest' }
      responses:
        '200':
          description: Updated assembly info with version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenericMap' }
  /split_contig_at_bin:
    post:
      tags: [Scaffolding]
      summary: Split contig at a given position
      operationId: splitContigAtBin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SplitContigRequest' }
      responses:
        '200':
          description: Updated assembly info with version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenericMap' }
  /group_contigs_into_scaffold:
    post:
      tags: [Scaffolding]
      summary: Group selected range into scaffold
      operationId: groupIntoScaffold
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupContigsRequest' }
      responses:
        '200':
          description: Updated assembly info with version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenericMap' }
  /ungroup_contigs_from_scaffold:
    post:
      tags: [Scaffolding]
      summary: Ungroup selected scaffold segment
      operationId: ungroupFromScaffold
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RangeRequest' }
      responses:
        '200':
          description: Updated assembly info with version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenericMap' }
  /move_selection_to_debris:
    post:
      tags: [Scaffolding]
      summary: Move selected range to debris
      operationId: moveSelectionToDebris
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RangeRequest' }
      responses:
        '200':
          description: Updated assembly info with version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenericMap' }
...
components:
...
  schemas:
...
    GenericMap:
      type: object
      additionalProperties: true
...
    RangeRequest:
      type: object
      properties:
        startBP: { type: integer, format: int64 }
        endBP: { type: integer, format: int64 }
      required: [startBP, endBP]
    MoveSelectionRequest:
      type: object
      properties:
        startBP: { type: integer, format: int64 }
        endBP: { type: integer, format: int64 }
        targetStartBP: { type: integer, format: int64 }
      required: [startBP, endBP, targetStartBP]
    SplitContigRequest:
      type: object
      properties:
        splitPx: { type: integer, format: int64 }
        bpResolution: { type: integer, format: int64 }
      required: [splitPx, bpResolution]
    GroupContigsRequest:
      type: object
      properties:
        startBP: { type: integer, format: int64 }
        endBP: { type: integer, format: int64 }
        newScaffoldName: { type: string }
        spacerLength: { type: integer, format: int64 }
      required: [startBP, endBP]
...